+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxow.hs
@@ -84,6 +84,7 @@ import Data.List.NonEmpty (NonEmpty)
 import qualified Data.Map.Strict as Map
 import Data.Set (Set)
 import qualified Data.Set as Set
+import Data.Set.NonEmpty (NonEmptySet)
 import Data.Typeable (Typeable)
 import GHC.Generics (Generic)
 import Lens.Micro
@@ -100,25 +101,20 @@ data AlonzoUtxowPredFailure era
     MissingRedeemers
       (NonEmpty (PlutusPurpose AsItem era, ScriptHash))
   | MissingRequiredDatums
-      -- TODO: Make this NonEmpty #4066
-
       -- | Set of missing data hashes
-      (Set DataHash)
+      (NonEmptySet DataHash)
       -- | Set of received data hashes
       (Set DataHash)
   | NotAllowedSupplementalDatums
-      -- TODO: Make this NonEmpty #4066
-
       -- | Set of unallowed data hashes
-      (Set DataHash)
+      (NonEmptySet DataHash)
       -- | Set of acceptable supplemental data hashes
       (Set DataHash)
   | PPViewHashesDontMatch
       (Mismatch RelEQ (StrictMaybe ScriptIntegrityHash))
   | -- | Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't
     UnspendableUTxONoDatumHash
-      -- TODO: Make this NonEmpty #4066
-      (Set TxIn)
+      (NonEmptySet TxIn)
   | -- | List of redeemers not needed
     ExtraRedeemers
       (NonEmpty (PlutusPurpose AsIx era))
@@ -246,15 +242,11 @@ missingRequiredDatums utxo tx = do
       (okSupplimentalDHs, notOkSupplimentalDHs) =
         Set.partition (`Set.member` allowedSupplementalDataHashes) supplimentalDatumHashes
   sequenceA_
-    [ failureUnless
-        (Set.null txInsNoDataHash)
-        (UnspendableUTxONoDatumHash txInsNoDataHash)
-    , failureUnless
-        (Set.null unmatchedDatumHashes)
-        (MissingRequiredDatums unmatchedDatumHashes txHashes)
-    , failureUnless
-        (Set.null notOkSupplimentalDHs)
-        (NotAllowedSupplementalDatums notOkSupplimentalDHs okSupplimentalDHs)
+    [ failureOnNonEmptySet txInsNoDataHash UnspendableUTxONoDatumHash
+    , failureOnNonEmptySet unmatchedDatumHashes (\unmatched -> MissingRequiredDatums unmatched txHashes)
+    , failureOnNonEmptySet
+        notOkSupplimentalDHs
+        (\notOk -> NotAllowedSupplementalDatums notOk okSupplimentalDHs)
     ]
 
 -- ==================