+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Imp/UtxowSpec.hs
@@ -20,6 +20,7 @@ import Cardano.Ledger.Shelley.Scripts (
   pattern RequireSignature,
  )
 import qualified Data.Set as Set
+import qualified Data.Set.NonEmpty as NES
 import Lens.Micro
 import Test.Cardano.Ledger.Core.KeyPair (ByronKeyPair (..))
 import Test.Cardano.Ledger.Imp.Common
@@ -62,7 +63,9 @@ spec = describe "UTXOW" $ do
       submitFailingTx
         tx
         [ injectFailure $
-            MissingVKeyWitnessesUTXOW [asWitness aliceKh]
+            MissingVKeyWitnessesUTXOW $
+              NES.singleton $
+                asWitness aliceKh
         ]
 
   it "MissingScriptWitnessesUTXOW" $ do
@@ -73,7 +76,8 @@ spec = describe "UTXOW" $ do
     submitFailingTx
       tx
       [ injectFailure $
-          MissingScriptWitnessesUTXOW [scriptHash]
+          MissingScriptWitnessesUTXOW $
+            NES.singleton scriptHash
       ]
 
   it "MissingTxBodyMetadataHash" $ do
@@ -125,9 +129,9 @@ spec = describe "UTXOW" $ do
       -- We dropped validating scripts when they are not needed, starting with Babbage
       if eraProtVerLow @era >= natVersion @6
         then
-          [ injectFailure $ ExtraneousScriptWitnessesUTXOW [scriptHash]
+          [ injectFailure $ ExtraneousScriptWitnessesUTXOW $ NES.singleton scriptHash
           ]
         else
-          [ injectFailure $ ScriptWitnessNotValidatingUTXOW [scriptHash]
-          , injectFailure $ ExtraneousScriptWitnessesUTXOW [scriptHash]
+          [ injectFailure $ ScriptWitnessNotValidatingUTXOW $ NES.singleton scriptHash
+          , injectFailure $ ExtraneousScriptWitnessesUTXOW $ NES.singleton scriptHash
           ]